+++ fe/src/hooks/useFunnel/type.ts
@@ -1,0 +1,16 @@
+import { ReactElement, ReactNode } from "react";
+
+export type NonEmptyArray<T> = readonly [T, ...T[]];
+
+export interface FunnelProps<Steps extends NonEmptyArray<string>> {
+  steps: Steps;
+  step: Steps[number];
+  children:
+    | Array<ReactElement<StepProps<Steps>>>
+    | ReactElement<StepProps<Steps>>;
+}
+
+export interface StepProps<Steps extends NonEmptyArray<string>> {
+  name: Steps[number];
+  children: ReactNode;
+}
